[ refactor ] 서버 전체 리팩토링
