+++ src/routes/authRoutes.ts
@@ -1,0 +1,7 @@
+import { Router } from "express";
+
+const router = Router();
+
+router.post("/login");
+
+export default router;
